move GetSizeInBytes to public FormatSizeHelpers (+ tests), make FormatHelpers internal again
