+++ src/Veldrid/MTL/MTLCommandList.cs
@@ -412,7 +412,7 @@ namespace Veldrid.MTL
                     uint compressedSrcX = srcX / blockSize;
                     uint compressedSrcY = srcY / blockSize;
                     uint blockSizeInBytes = blockSize == 1
-                        ? FormatHelpers.GetSizeInBytes(srcMTLTexture.Format)
+                        ? FormatSizeHelpers.GetSizeInBytes(srcMTLTexture.Format)
                         : FormatHelpers.GetBlockSizeInBytes(srcMTLTexture.Format);
 
                     ulong srcSubresourceBase = Util.ComputeSubresourceOffset(
@@ -482,7 +482,7 @@ namespace Veldrid.MTL
                     uint blockSize = FormatHelpers.IsCompressedFormat(dstMTLTexture.Format) ? 4u : 1u;
                     if (blockSize == 1)
                     {
-                        uint pixelSize = FormatHelpers.GetSizeInBytes(dstMTLTexture.Format);
+                        uint pixelSize = FormatSizeHelpers.GetSizeInBytes(dstMTLTexture.Format);
                         uint copySize = width * pixelSize;
                         for (uint zz = 0; zz < depth; zz++)
                             for (uint yy = 0; yy < height; yy++)
@@ -557,7 +557,7 @@ namespace Veldrid.MTL
                     uint compressedDstX = dstX / blockSize;
                     uint compressedDstY = dstY / blockSize;
                     uint blockSizeInBytes = blockSize == 1
-                        ? FormatHelpers.GetSizeInBytes(srcMTLTexture.Format)
+                        ? FormatSizeHelpers.GetSizeInBytes(srcMTLTexture.Format)
                         : FormatHelpers.GetBlockSizeInBytes(srcMTLTexture.Format);
                     uint rowPitch = FormatHelpers.GetRowPitch(bufferRowLength, srcMTLTexture.Format);
                     uint depthPitch = FormatHelpers.GetDepthPitch(rowPitch, bufferImageHeight, srcMTLTexture.Format);
