+++ src/Veldrid/OpenGL/OpenGLTexture.cs
@@ -661,7 +661,7 @@ namespace Veldrid.OpenGL
                 glBindBuffer(BufferTarget.CopyWriteBuffer, _pbos[subresource]);
                 CheckLastError();
 
-                uint dataSize = Width * Height * FormatHelpers.GetSizeInBytes(Format);
+                uint dataSize = Width * Height * FormatSizeHelpers.GetSizeInBytes(Format);
                 glBufferData(
                     BufferTarget.CopyWriteBuffer,
                     (UIntPtr)dataSize,
